data-generators/requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pymongo==4.6.0
faker==20.1.0